Adding full set of Sept. 2022 transcriptions by S. Guillaume and G. Wisniewski
Making public a set of transcriptions produced by Séverine Guillaume and Guillaume Wisniewski. They created an acoustic model for Na by using the (hand-made) transcriptions of Na texts in the present repository as a training corpus. Then this acoustic model was used for transcribing fresh materials.
The documents are straight from the model, without corrections.
An XML file, metadata.xml, can be used to identify the relevant documents in the Pangloss Collection using the number assigned in file names.